Changed component scanning path. Tidied up ForgerockOpenbankingGatewayApplication (#210)
