+++ forgerock-openbanking-aspsp/forgerock-openbanking-aspsp-as/forgerock-openbanking-as-api/src/main/resources/filtered/bootstrap.properties
@@ -2,3 +2,4 @@ spring.application.name=@project.name@
 spring.cloud.config.fail-fast=true
 spring.cloud.config.retry.max-interval=30000
 spring.cloud.config.retry.max-attempts=30
+spring.main.allow-bean-definition-overriding=true
