+++ forgerock-openbanking-backstage/forgerock-openbanking-admin/src/main/java/com/forgerock/openbanking/forgerockopenbankingadmin/ForgerockOpenbankingAdminApplication.java
@@ -26,7 +26,7 @@ import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.cloud.client.discovery.EnableDiscoveryClient;
 import org.springframework.context.annotation.Import;
 
-@SpringBootApplication(scanBasePackages = "com.forgerock")
+@SpringBootApplication(scanBasePackages = "com.forgerock.openbanking")
 @EnableAdminServer
 @EnableDiscoveryClient
 @Import(AdminApplicationSecurityConfiguration.class)
