+++ forgerock-openbanking-devportal/forgerock-openbanking-docs/src/main/java/com/forgerock/openbanking/docs/ForgerockOpenbankingDocsApplication.java
@@ -27,7 +27,7 @@ import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.cloud.client.discovery.EnableDiscoveryClient;
 import org.springframework.context.annotation.Import;
 
-@SpringBootApplication(scanBasePackages = "com.forgerock")
+@SpringBootApplication(scanBasePackages = "com.forgerock.openbanking")
 @EnableDiscoveryClient
 @EnableSslClient
 @Import(AuthWebSecurityConfiguration.class)
